poc: common JSX components
