+++ packages/instantsearch-jsx/tsconfig.declaration.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../tsconfig.declaration"
+}
